Update data-routing.md
Refer to https://github.com/couchbase/couchbase-lite-ios/issues/29 for details.  Also, from a chat post by Jens: It’s similar to a deletion in that there’s a special tombstone revision, it’s just that the revision has `_removed:true` instead of `_deleted:true`. This revision doesn’t really exist in SG, at least not in that form; it’s synthesized so that CBL will no longer have the previous revision of the doc.